feat: better modal types
