+++ packages/modal/Modal.stories.tsx
@@ -1,15 +1,12 @@
+import { useCallback, useState } from 'react'
 import styled from 'styled-components'
 import { Story, Meta } from '@storybook/react'
-import { ModalProps } from './types'
 import { Button, ButtonProps, ButtonSize } from '@lidofinance/button'
 import { Loader } from '@lidofinance/loaders'
 import { Text } from '@lidofinance/text'
 import { Link } from '@lidofinance/link'
 import { Error, Success, Eth, Terra } from '@lidofinance/icons'
-import Modal from './Modal'
-import ModalExtra from './ModalExtra'
-import ModalButton from './ModalButton'
-import { useCallback, useState } from 'react'
+import { Modal, ModalExtra, ModalButton, ModalProps } from '.'
 
 const getOptions = (enumObject: Record<string, string | number>) =>
   Object.values(enumObject).filter((value) => typeof value === 'string')
@@ -31,8 +28,7 @@ export default {
   },
 } as Meta
 
-const useModal = (props: ModalProps) => {
-  const { onClose, onBack } = props
+const useModal = ({ onClose, onBack }: ModalProps) => {
   const [state, setState] = useState(false)
   const handleOpen = useCallback(() => setState(true), [])
   const handleClose = useCallback(() => {
