+++ packages/modal/ModalExtraStyles.tsx
@@ -1,4 +1,7 @@
+import { InferStyledComponentPropsWithoutRef } from '@lidofinance/utils'
 import styled from 'styled-components'
+
+export type ModalExtraStyleProps = InferStyledComponentPropsWithoutRef<typeof ModalExtraStyle>
 
 export const ModalExtraStyle = styled.div`
   margin-top: ${({ theme }) => -theme.borderRadiusesMap.xl}px;
