+++ packages/modal/index.ts
@@ -1,5 +1,8 @@
 export { default as Modal } from './Modal'
+export * from './Modal'
 export { default as ModalExtra } from './ModalExtra'
+export * from './ModalExtra'
 export { default as ModalOverlay } from './ModalOverlay'
+export * from './ModalOverlay'
 export { default as ModalButton } from './ModalButton'
-export * from './types'
+export * from './ModalButton'
